rename: comments 관련 파일  vote 하위 디렉터리로 이동
