+++ src/feature/votes/votes.module.ts
@@ -8,8 +8,8 @@ import { CommentsModule } from "./vote/comments/comments.module";
  * Vote module
  */
 @Module({
-  imports: [VoteModule, CommentsModule],
-  controllers: [VotesController],
-  providers: [VotesService],
+    imports: [ VoteModule, CommentsModule ],
+    controllers: [ VotesController ],
+    providers: [ VotesService ]
 })
 export class VotesModule {}
